Lab 7: Read the controller
Wow!  Real player control!  In this lab, we cover how to strobe and read
the dpad buttons and hook it in to the existing "engine", such as it is,
so that the little character runs forward only when controlled to do so.

The process is easy, involving strobing the button state into a shift
register and reading it back.  There are some pretty standard algorithms
for doing this, but again, the goal of many of these labs is to make
something as plain and readable as possible first.